requirements.txt
sentence-transformers
numpy
weaviate-client
PyMuPDF
